bank service. Problem with super class FK.

+++ src/main/java/com/bank/dtos/CurrentBankDto.java
@@ -2,7 +2,6 @@ package com.bank.dtos;
 
 import com.bank.enums.AccountStatus;
 import lombok.Data;
-
 import java.io.Serializable;
 import java.util.Date;
 
@@ -10,9 +9,9 @@ import java.util.Date;
 public class CurrentBankDto extends  BankAccountDto implements Serializable {
 
     private String id;
-    private Double balance;
+    private Float balance;
     private Date createAt;
     private AccountStatus status;
     private CustomerDto customerDto;
-    private Double overDraft;
+    private Float overDraft;
 }
